refactor: social-filter-bar폴더 구조 컨벤션대로 변경
